requirements.txt
python-binance==1.0.17
requests==2.31.0
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
python-dotenv==1.0.0
ccxt==4.0.0
beautifulsoup4==4.12.2
tweepy==4.14.0
textblob==0.17.1
ta==0.10.2
numpy==1.24.3
aiohttp==3.8.5
python-telegram-bot==20.4
praw==7.7.1
